api/requirements.txt
djangorestframework
django-filter
django-cors-headers
django-environ
# django_neomodel
# neomodel
psycopg2-binary
markdown

regex
requests
logging
urllib3


networkx
gunicorn

celery
matplotlib
